# backend/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "VitaNote API",
    "version": "1.0.0",
    "description": "VitaNote - Smart Health Management API",
    "contact": {
      "name": "VitaNote Team",
      "email": "support@vitanote.com"
    }
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "User registration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful registration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "User login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["userName", "email", "password"],
        "properties": {
          "userName": {"type": "string"},
          "email": {"type": "string"},
          "password": {"type": "string"},
          "phoneNumber": {"type": "string"}
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["userName", "password"],
        "properties": {
          "userName": {"type": "string"},
          "password": {"type": "string"}
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {"type": "string"},
          "refreshToken": {"type": "string"},
          "tokenExpires": {"type": "string", "format": "date-time"},
          "user": {"$ref": "#/components/schemas/UserData"}
        }
      },
      "UserData": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "userName": {"type": "string"},
          "email": {"type": "string"},
          "phoneNumber": {"type": "string"}
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}